Refactor {RelatedItems,ExternalResources}EditView (#455)
