+++ frontend/src/forms/add-button-template.hbs
@@ -1,0 +1,4 @@
+<button type=button class="button">
+    <span>Add</span>
+    <span class="icon is-small"><i class="fas fa-plus"></i></span>
+</button>
