+++ frontend/src/forms/multifield-template.hbs
@@ -1,0 +1,1 @@
+<div class="field rit-add-row"></div>
