+++ frontend/src/panel-common/field-editing-template.hbs
@@ -1,20 +1,9 @@
 <div class="panel-header">
     <header>
-        <h3 class="title is-3">Related items</h3>
+        <h3 class="title is-3">{{title}}</h3>
     </header>
 </div>
-<div class="panel-content relations">
-    <div class="field rit-add-relation">
-        <div class="control">
-            <button type=button class="button">
-                <span>Add</span>
-                <span class="icon is-small">
-                    <i class="fas fa-plus"></i>
-                </span>
-            </button>
-        </div>
-    </div>
-</div>
+{{! multifield inserted here }}
 <div class="panel-footer field is-grouped is-grouped-left">
     <div class="control">
         <button
