--- frontend/src/panel-external-resources/external-resources-edit-template.hbs
@@ -1,32 +1,0 @@
-<div class="panel-header">
-    <header>
-        <h3 class="title is-3">Edit external resources</h3>
-    </header>
-</div>
-<div class="panel-content external">
-   {{! external resources of ExternalResourceEditItem rendered here }} 
-    <div class="field rit-add-external">
-        <div class="control">
-            <button type=button class="button add">
-            <span>Add</span>
-            <span class="icon is-small">
-                <i class="fas fa-plus"></i>
-            </span>
-            </button>
-        </div>
-    </div>
-</div>
-<div class="panel-footer field is-grouped is-grouped-left">
-    <div class="control">
-        <button
-            type="submit"
-            class="button btn-save submit"
-        ><span>Save</span></button>
-    </div>
-    <div class="control">
-        <button
-            type="reset"
-            class="button btn-cancel is-primary"
-        ><span>Cancel</span></button>
-    </div>
-</div>
