+++ frontend/src/panel-external-resources/external-resources-edit-view-test.ts
@@ -19,19 +19,22 @@ describe('ExternalResourceEditItemView', function() {
 
     it('registers changes of complete models', function() {
         this.changedModel.set('object', 'and another thing');
-        this.view.updateExternalResource(this.changedModel);
-        expect(this.view.changes.models[0].get('action')).toEqual('set');
+        const rowManager = this.view.rowManager;
+        rowManager.updateExternalResource(this.changedModel);
+        expect(rowManager.changes.models[0].get('action')).toEqual('set');
     })
 
     it('does not register changes of incomplete models', function() {
-        this.view.updateExternalResource(this.changedModel);
-        expect(this.view.changes.models.length).toEqual(0);
+        const rowManager = this.view.rowManager;
+        rowManager.updateExternalResource(this.changedModel);
+        expect(rowManager.changes.models.length).toEqual(0);
     })
 
     it('does not register changes if predicate is not defined', function() {
         this.changedModel.set('predicate', undefined);
         this.changedModel.set('object', 'and another thing');
-        this.view.updateExternalResource(this.changedModel);
-        expect(this.view.changes.models.length).toEqual(0);
+        const rowManager = this.view.rowManager;
+        rowManager.updateExternalResource(this.changedModel);
+        expect(rowManager.changes.models.length).toEqual(0);
     })
 });
